dom/workers/test/test_errorPropagation.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
  <head>
    <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js">
    </script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  </head>
  <body>
    <iframe id="workerFrame" src="errorPropagation_iframe.html"
            onload="workerFrameLoaded();"></iframe>
    <script type="text/javascript">
      const workerCount = 3;

      const errorMessage = "expectedError";
      const errorFilename = "http://mochi.test:8888/tests/dom/workers/test/" +
                            "errorPropagation_worker.js";
      const errorLineno = 48;

      var workerFrame;

      scopeErrorCount = 0;
      workerErrorCount = 0;
      windowErrorCount = 0;

      function messageListener(event) {
        if (event.type == "scope") {
          scopeErrorCount++;
        }
        else if (event.type == "worker") {
          workerErrorCount++;
        }
        else if (event.type == "window") {
          windowErrorCount++;
        }
        else {
          ok(false, "Bad event type: " + event.type);
        }

        is(event.data.message, errorMessage, "Correct message event.message");
        is(event.data.filename, errorFilename,
           "Correct message event.filename");
        is(event.data.lineno, errorLineno, "Correct message event.lineno");
      }

      function consoleListener(message) {
        is(message,
           "[JavaScript Error: \"" + errorMessage + "\" {file: \"" +
           errorFilename + "\" line: " + errorLineno + "}]",
           "Correct message in error console");

        is(scopeErrorCount, workerCount, "Good number of scope errors");
        is(workerErrorCount, workerCount, "Good number of worker errors");
        is(windowErrorCount, 1, "Good number of window errors");

        workerFrame.stop();
        SpecialPowers.removeErrorConsoleListener(consoleListener);
        SimpleTest.finish();
      }

      function workerFrameLoaded() {
        workerFrame = document.getElementById("workerFrame").contentWindow;
        SpecialPowers.addErrorConsoleListener(consoleListener);
        workerFrame.start(workerCount, messageListener);
      }

      SimpleTest.waitForExplicitFinish();
    </script>
  </body>
</html>
